# openapi/ars-api.yaml
openapi: "3.1.0"

info:
  title: SmartRetail Aggregated Reporting Service API
  version: "1.0.0"
  description: |
    Read-only aggregation service that assembles dashboard payloads for each persona.
    ARS queries the forecasting, inventory, replenishment, and supplier schemas
    using parallel queries merged in Java — no cross-schema SQL JOINs.
    Role-based access is enforced per endpoint.

    ARS is responsible for:
      /v1/dashboard/*          — persona-specific aggregated KPI payloads

    Inventory alerts/positions  → IMS  (port 8081, ims-api.yaml)
    Replenishment orders (CRUD) → RE   (port 8082, re-api.yaml)
    Demand forecast bands       → DFS  (port 8084, dfs-api.yaml)
    Supplier order tracking     → SUP  (port 8085, sup-api.yaml)
  contact:
    name: SmartRetail Platform Team

servers:
  - url: http://localhost:8083
    description: Local development
  - url: https://{api-id}.execute-api.us-east-1.amazonaws.com/internal
    description: AWS internal stage

tags:
  - name: dashboard
    description: Persona-specific aggregated dashboard payloads

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      $ref: './components/schemas.yaml#/ErrorResponse'

    # ── Shared enums ────────────────────────────────────────────────────────────

    Trend:
      type: string
      description: Direction of movement vs the prior comparison period
      enum:
        - IMPROVING
        - STABLE
        - DEGRADING
      example: IMPROVING

    DirectionTrend:
      type: string
      description: Direction of movement for count-based KPIs
      enum:
        - INCREASING
        - STABLE
        - DECREASING
      example: DECREASING

    # ── Executive KPI schemas (Flow 8) ──────────────────────────────────────────

    MapeDataPoint:
      type: object
      description: MAPE value for a single completed forecast run
      required:
        - runDate
        - mape
      additionalProperties: false
      properties:
        runDate:
          type: string
          format: date
          description: Calendar date of the forecast run (UTC)
          example: "2026-05-15"
        mape:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Mean Absolute Percentage Error for this run (0.0–1.0 scale)
          example: 0.0823

    ForecastAccuracyKpi:
      type: object
      description: Forecast accuracy derived from the most recent completed forecast run
      required:
        - latestMape
        - trend
        - history
      additionalProperties: false
      properties:
        latestMape:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: MAPE of the most recent COMPLETED forecast run
          example: 0.0823
        trend:
          $ref: '#/components/schemas/Trend'
        history:
          type: array
          description: MAPE values for the last 30 completed forecast runs, newest first
          items:
            $ref: '#/components/schemas/MapeDataPoint'

    StockoutAlertDataPoint:
      type: object
      description: Daily count of CRITICAL stock alerts for the Stockout Frequency chart
      required:
        - alertDate
        - criticalCount
      additionalProperties: false
      properties:
        alertDate:
          type: string
          format: date
          description: Calendar date (UTC) on which these alerts were raised
          example: "2026-05-15"
        criticalCount:
          type: integer
          format: int32
          minimum: 0
          description: Number of CRITICAL alerts raised on this date
          example: 3

    CycleTimeDataPoint:
      type: object
      description: Weekly average replenishment cycle time for the Cycle Time chart
      required:
        - weekStart
        - averageDays
        - poCount
      additionalProperties: false
      properties:
        weekStart:
          type: string
          format: date
          description: Monday of the ISO week (UTC)
          example: "2026-05-12"
        averageDays:
          type: number
          format: double
          minimum: 0
          description: Average calendar days from PO creation to DISPATCHED/COMPLETED in this week
          example: 3.4
        poCount:
          type: integer
          format: int32
          minimum: 0
          description: Number of completed purchase orders in this week
          example: 5

    StockoutFrequencyKpi:
      type: object
      description: Count of CRITICAL stock alerts in the last 30 days vs the prior 30-day window
      required:
        - last30Days
        - trend
        - history
      additionalProperties: false
      properties:
        last30Days:
          type: integer
          format: int32
          minimum: 0
          description: Number of CRITICAL stock alerts raised in the last 30 days
          example: 12
        trend:
          $ref: '#/components/schemas/DirectionTrend'
        history:
          type: array
          description: Daily CRITICAL alert counts for the last 30 days, newest first
          items:
            $ref: '#/components/schemas/StockoutAlertDataPoint'

    ReplenishmentCycleTimeKpi:
      type: object
      description: Average calendar days from purchase order creation to DISPATCHED/COMPLETED status
      required:
        - averageDays
        - trend
        - history
      additionalProperties: false
      properties:
        averageDays:
          type: number
          format: double
          minimum: 0
          description: Average days from PO creation to DISPATCHED/COMPLETED over the last 90 days
          example: 3.2
        trend:
          $ref: '#/components/schemas/Trend'
        history:
          type: array
          description: Weekly average cycle time for the last 90 days, newest first
          items:
            $ref: '#/components/schemas/CycleTimeDataPoint'

    OnTimeDeliveryKpi:
      type: object
      description: Aggregate on-time delivery rate across all active suppliers
      required:
        - rate
        - trend
      additionalProperties: false
      properties:
        rate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Aggregate OTD rate across all suppliers (0.0–1.0)
          example: 0.848
        trend:
          $ref: '#/components/schemas/Trend'

    SupplierPerformanceEntry:
      type: object
      description: Delivery performance metrics for a single supplier (Executive view)
      required:
        - supplierId
        - supplierName
        - otdRate
        - fillRate
        - earlyCount
        - onTimeCount
        - lateCount
        - openExceptions
      additionalProperties: false
      properties:
        supplierId:
          type: string
          format: uuid
          example: "11111111-0000-0000-0000-000000000001"
        supplierName:
          type: string
          example: "Acme Beverages Ltd"
        otdRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.92
        fillRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 1.0
        earlyCount:
          type: integer
          format: int32
          minimum: 0
          example: 2
        onTimeCount:
          type: integer
          format: int32
          minimum: 0
          example: 7
        lateCount:
          type: integer
          format: int32
          minimum: 0
          example: 1
        openExceptions:
          type: integer
          format: int32
          minimum: 0
          example: 1

    ExecutiveKpis:
      type: object
      required:
        - forecastAccuracy
        - stockoutFrequency
        - replenishmentCycleTime
        - onTimeDelivery
        - supplierPerformance
      additionalProperties: false
      properties:
        forecastAccuracy:
          $ref: '#/components/schemas/ForecastAccuracyKpi'
        stockoutFrequency:
          $ref: '#/components/schemas/StockoutFrequencyKpi'
        replenishmentCycleTime:
          $ref: '#/components/schemas/ReplenishmentCycleTimeKpi'
        onTimeDelivery:
          $ref: '#/components/schemas/OnTimeDeliveryKpi'
        supplierPerformance:
          type: array
          description: Per-supplier performance metrics, sorted by OTD rate descending
          items:
            $ref: '#/components/schemas/SupplierPerformanceEntry'

    ExecutiveDashboardResponse:
      type: object
      description: |
        Executive Insights Dashboard payload.
        Data is assembled from parallel reads across the forecasting, inventory,
        and replenishment schemas — never via cross-schema SQL joins.
      required:
        - kpis
        - dataFreshness
      additionalProperties: false
      properties:
        kpis:
          $ref: '#/components/schemas/ExecutiveKpis'
        dataFreshness:
          type: string
          format: date-time
          description: Earliest completed_at / created_at timestamp across all contributing queries
          example: "2026-05-15T14:23:00Z"

    # ── SC Planner Dashboard schemas (Flow 9) ───────────────────────────────────

    ScPlannerForecastAccuracy:
      type: object
      description: Compact forecast accuracy summary for SC Planner dashboard header
      required:
        - latestMape
        - mapeThreshold
        - lastRunAt
        - status
      additionalProperties: false
      properties:
        latestMape:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.0823
        mapeThreshold:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.15
        lastRunAt:
          type: string
          format: date-time
          example: "2026-05-15T02:00:00Z"
        status:
          type: string
          enum:
            - WITHIN_THRESHOLD
            - ABOVE_THRESHOLD
          example: WITHIN_THRESHOLD

    ScPlannerDashboardResponse:
      type: object
      description: |
        SC Planner dashboard summary — drives tab badges and header metrics.
        Parallel reads from inventory.stock_alerts, replenishment.purchase_orders,
        and forecasting.forecast_runs — no cross-schema SQL joins.
      required:
        - pendingApprovalCount
        - activeAlertCount
        - forecastAccuracy
        - dataFreshness
      additionalProperties: false
      properties:
        pendingApprovalCount:
          type: integer
          format: int32
          minimum: 0
          description: POs in PENDING_APPROVAL status
          example: 3
        activeAlertCount:
          type: integer
          format: int32
          minimum: 0
          description: Stock alerts in ACTIVE status
          example: 43
        forecastAccuracy:
          $ref: '#/components/schemas/ScPlannerForecastAccuracy'
        dataFreshness:
          type: string
          format: date-time
          example: "2026-05-15T14:23:00Z"

    # ── Supplier Performance Scorecard schemas (Flow 9) ─────────────────────────

    ScPlannerSupplierEntry:
      type: object
      description: Full supplier performance metrics for SC Planner scorecard
      required:
        - supplierId
        - supplierName
        - onTimeDeliveryRate
        - poAcknowledgementSlaCompliance
        - openExceptions
        - avgLeadTimeVarianceDays
        - totalPoCount
        - totalPoValue
      additionalProperties: false
      properties:
        supplierId:
          type: string
          format: uuid
          example: "11111111-0000-0000-0000-000000000001"
        supplierName:
          type: string
          example: "Acme Beverages Ltd"
        onTimeDeliveryRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Dispatched within lead_time_days of approval
          example: 0.92
        poAcknowledgementSlaCompliance:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.92
        openExceptions:
          type: integer
          format: int32
          minimum: 0
          example: 2
        avgLeadTimeVarianceDays:
          type: number
          format: double
          description: Positive = late, negative = early
          example: 1.3
        totalPoCount:
          type: integer
          format: int32
          minimum: 0
          example: 45
        totalPoValue:
          type: number
          format: double
          minimum: 0
          description: Total GBP value of POs in the last 90 days
          example: 234500.00

    SupplierPerformanceDashboardResponse:
      type: object
      description: |
        Supplier performance scorecard for SC Planner.
        Three separate queries (supplier names, replenishment PO metrics, shipment metrics)
        merged in Java by supplierId — no cross-schema SQL joins.
      required:
        - suppliers
        - dataFreshness
      additionalProperties: false
      properties:
        suppliers:
          type: array
          description: Sorted by onTimeDeliveryRate ascending (worst first)
          items:
            $ref: '#/components/schemas/ScPlannerSupplierEntry'
        dataFreshness:
          type: string
          format: date-time
          example: "2026-05-15T14:23:00Z"

  responses:
    Unauthorized:
      $ref: './components/responses.yaml#/Unauthorized'
    Forbidden:
      $ref: './components/responses.yaml#/Forbidden'
    InternalServerError:
      $ref: './components/responses.yaml#/InternalServerError'

paths:

  # ── Executive Dashboard (Flow 8) ─────────────────────────────────────────────

  /v1/dashboard/executive:
    get:
      tags:
        - dashboard
      operationId: getExecutiveDashboard
      summary: Executive Insights Dashboard
      description: |
        Returns MAPE trend, stockout frequency, and replenishment cycle time KPIs.
        Allowed roles: EXECUTIVE, SC_PLANNER, ADMIN.
        Data sources (parallel queries, merged in Java):
        - forecasting.forecast_runs — last 30 COMPLETED runs
        - inventory.stock_alerts — CRITICAL alert counts
        - replenishment.purchase_orders — average cycle time, last 90 days
      responses:
        '200':
          description: Executive dashboard KPIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutiveDashboardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── Store Manager Dashboard (Flow 4) ─────────────────────────────────────────

  /v1/dashboard/store-manager:
    get:
      tags:
        - dashboard
      operationId: getStoreManagerDashboard
      summary: Store Manager Dashboard (Flow 4 — not yet implemented)
      description: |
        Allowed roles: STORE_MANAGER, ADMIN.
        Implementation planned for Flow 4.
      parameters:
        - name: dcId
          in: query
          required: true
          schema:
            type: string
            maxLength: 50
          example: "DC-LONDON"
      responses:
        '200':
          description: Store Manager dashboard payload
          content:
            application/json:
              schema:
                type: object
                description: Store Manager dashboard — schema defined in Flow 4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '501':
          description: Not yet implemented — available in Flow 4
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── SC Planner Dashboard (Flow 9) ────────────────────────────────────────────

  /v1/dashboard/sc-planner:
    get:
      tags:
        - dashboard
      operationId: getScPlannerDashboard
      summary: SC Planner Dashboard Summary
      description: |
        Returns pending approval count, active alert count, and forecast accuracy
        for the SC Planner Console header and tab badges (Flow 9).
        Allowed roles: SC_PLANNER, ADMIN.
        Parallel reads — no cross-schema SQL joins:
        - replenishment.purchase_orders: COUNT PENDING_APPROVAL
        - inventory.stock_alerts: COUNT ACTIVE
        - forecasting.forecast_runs: latest MAPE
      responses:
        '200':
          description: SC Planner dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScPlannerDashboardResponse'
              example:
                pendingApprovalCount: 3
                activeAlertCount: 43
                forecastAccuracy:
                  latestMape: 0.0823
                  mapeThreshold: 0.15
                  lastRunAt: "2026-05-15T02:00:00Z"
                  status: WITHIN_THRESHOLD
                dataFreshness: "2026-05-15T14:23:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── Supplier Performance Scorecard (Flow 9) ──────────────────────────────────

  /v1/dashboard/supplier-performance:
    get:
      tags:
        - dashboard
      operationId: getSupplierPerformanceDashboard
      summary: Supplier Performance Scorecard
      description: |
        Returns OTD rate, SLA compliance, open exceptions, and lead time variance
        per supplier for the SC Planner Console scorecard tab (Flow 9).
        Allowed roles: SC_PLANNER, ADMIN.
        Three separate queries merged in Java by supplierId — no cross-schema SQL joins:
        - supplier.supplier_records: active supplier names
        - replenishment.purchase_orders: PO counts and values (last 90 days)
        - supplier.supplier_pos + supplier.shipment_updates: OTD metrics
      responses:
        '200':
          description: Supplier performance scorecard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierPerformanceDashboardResponse'
              example:
                suppliers:
                  - supplierId: "11111111-0000-0000-0000-000000000005"
                    supplierName: "Metro Food Distributors"
                    onTimeDeliveryRate: 0.71
                    poAcknowledgementSlaCompliance: 0.80
                    openExceptions: 3
                    avgLeadTimeVarianceDays: 2.1
                    totalPoCount: 10
                    totalPoValue: 85000.00
                  - supplierId: "11111111-0000-0000-0000-000000000004"
                    supplierName: "Chill Chain Logistics"
                    onTimeDeliveryRate: 0.95
                    poAcknowledgementSlaCompliance: 0.98
                    openExceptions: 0
                    avgLeadTimeVarianceDays: -0.3
                    totalPoCount: 10
                    totalPoValue: 120000.00
                dataFreshness: "2026-05-15T14:23:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'